refactor(aigc-video-gen): aigc_common.py 移到 scripts/ 与三脚本同目录
原放 skills/_shared/ 需 sys.path.insert 跨目录跳转，import 不干净。
aigc_common.py 只被本 skill 的三个 gen_*.py 调用，放 scripts/ 同级更合理：
- 三个 gen_*.py 的 sys.path.insert 改为指向同目录
- SKILL.md 引用从 skills/_shared/aigc_common.py 改为 scripts/aigc_common.py
- 旧 skills/_shared/aigc_common.py 清空待删

验证：py_compile 四文件 OK + sh dispatch 四测试符合预期
（--platform minimax / env 自动判 / ref-audio 约束 / 全无 env exit 2）
