+++ skills/aigc-video-gen/scripts/gen_minimax.py
@@ -32,8 +32,8 @@ import urllib.error
 import urllib.request
 from pathlib import Path
 
-# _shared/aigc_common.py 在 skills/_shared/，本脚本在 skills/aigc-video-gen/scripts/
-sys.path.insert(0, str(Path(__file__).resolve().parent.parent.parent / "_shared"))
+# aigc_common.py 与本脚本同目录（skills/aigc-video-gen/scripts/）
+sys.path.insert(0, str(Path(__file__).resolve().parent))
 from aigc_common import (  # noqa: E402
     HttpError,
     TaskFailed,
